Move /circuits endpoints to libsplinter
Move the /circuits endpoints from splinterd to libsplinter. Also add the
experimental `circuit-read` feature to libsplinter, which is required by
the splinter dameon's `circuit-read` feature.
