+++ libsplinter/Cargo.toml
@@ -77,6 +77,7 @@ experimental = [
     "biome",
     "biome-credentials",
     "biome-notifications",
+    "circuit-read",
     "connection-manager",
     "database",
     "zmq-transport",
@@ -85,6 +86,7 @@ experimental = [
 biome = ["database", "jsonwebtoken", "rand"]
 biome-credentials = ["biome", "database", "bcrypt"]
 biome-notifications = ["biome", "database"]
+circuit-read = []
 connection-manager = []
 database = ["diesel", "diesel_migrations"]
 events = ["hyper", "tokio", "awc"]
