+++ libsplinter/src/circuit/mod.rs
@@ -11,8 +11,10 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 pub mod directory;
 pub mod handlers;
+pub mod rest_api;
 pub mod service;
 
 use serde_derive::{Deserialize, Serialize};
